Added csvReader and functions to classify a star based off of the csv.

+++ affinityScoreCalculator.cpp
@@ -1,4 +1,5 @@
-int affinityScoreCalculator(int resourceValueModifier, int habitabilityModifier)
+#include <cstdint>
+int affinityScoreCalculator(int8_t resourceValueModifier, int8_t habitabilityModifier)
 {
 	int affinityScore;
 	affinityScore = resourceValueModifier + habitabilityModifier;
