+++ declarations/functions/affinityScoreCalculator.h
@@ -1,1 +1,2 @@
-int affinityScoreCalculator(int resourceValueModifier, int habitabilityModifier);
+#include <cstdint>
+int affinityScoreCalculator(int8_t resourceValueModifier, int8_t habitabilityModifier);
