+++ declarations/functions/habitabilityModifiersBreathableClimate.h
@@ -1,1 +1,2 @@
-int habitabilityModifiersBreathableClimate(char climateType, bool breathable);
+#include <cstdint>
+int8_t habitabilityModifiersBreathableClimate(char climateType, bool breathable);
