+++ declarations/functions/habitabilityModifiersNonBreathable.h
@@ -1,1 +1,3 @@
-int habitabilityModifiersNonBreathable(char pressureCategory, atmosphericComposition_t atmosphereComposition);
+#include <cstdint>
+#include "../structures/atmosphericComposition_t.h"
+int8_t habitabilityModifiersNonBreathable(char pressureCategory, atmosphericComposition_t atmosphereComposition);
