+++ declarations/functions/habitabilityModifiersTable.h
@@ -1,1 +1,3 @@
-int habitabilityModifiersTable(char worldType, float atmosphereMass, float hydrographicCoverage, float atmosphericPressure, int surfaceTemperature, atmosphericComposition_t atmosphereCompositionArray);
+#include <cstdint>
+#include "../structures/atmosphericComposition_t.h"
+int8_t habitabilityModifiersTable(char worldType, float atmosphereMass, float hydrographicCoverage, float atmosphericPressure, int surfaceTemperature, atmosphericComposition_t atmosphereCompositionArray);
