+++ declarations/functions/placeMoonOrbits.h
@@ -1,2 +1,3 @@
+#include <cstdint>
 //	This function creates orbital slots for a planet's major moons
-float placeMoonOrbits(float planetDiameter, int moonSizeClass, int family);
+float placeMoonOrbits(float planetDiameter, int8_t moonSizeClass, int8_t family);
