+++ habitabilityModifiersBreathable.cpp
@@ -1,9 +1,10 @@
 //	Constant declarations
+#include <cstdint>
 #include "declarations/constants/atmosphericPressureCategoryConstants.h"
 
-int habitabilityModifiersBreathable(char pressureCategory, bool breathable)
+int8_t habitabilityModifiersBreathable(char pressureCategory, bool breathable)
 {
-	int habitabilityModifier;
+	int8_t habitabilityModifier;
 
 	if 		(breathable == false																				) {habitabilityModifier = 0;}
 	else if (breathable == true && pressureCategory == APC_VERY_THIN											) {habitabilityModifier = 1;}
