+++ habitabilityModifiersLiquidWaterOceans.cpp
@@ -1,6 +1,7 @@
-int habitabilityModfiersLiquidWaterOceans(float hydrographicCoverage)
+#include <cstdint>
+int8_t habitabilityModfiersLiquidWaterOceans(float hydrographicCoverage)
 {
-	int habitabilityModifier;
+	int8_t habitabilityModifier;
 
 	if 		(hydrographicCoverage >= 0.01 && hydrographicCoverage <= 0.59) {habitabilityModifier = 1;}
 	else if (hydrographicCoverage >= 0.6  && hydrographicCoverage <= 0.9 ) {habitabilityModifier = 2;}
