+++ habitabilityModifiersNonBreathable.cpp
@@ -1,3 +1,4 @@
+#include <cstdint>
 //	Constant declarations
 #include "declarations/constants/atmosphericPressureCategoryConstants.h"
 
@@ -6,9 +7,9 @@
 
 using namespace std;
 
-int habitabilityModifiersNonBreathable(char pressureCategory, atmosphericComposition_t atmosphereComposition)
+int8_t habitabilityModifiersNonBreathable(char pressureCategory, atmosphericComposition_t atmosphereComposition)
 {
-	int habitabilityModifier = 0;
+	int8_t habitabilityModifier = 0;
 
 	if 		((pressureCategory == APC_VERY_THIN || pressureCategory == APC_TRACE) && (atmosphereComposition.toxicity > 0) && (atmosphereComposition.suffocating == 1) && (atmosphereComposition.corrosive == 1)) 	{habitabilityModifier = -2;}
 	else if ((pressureCategory == APC_VERY_THIN || pressureCategory == APC_TRACE) && (atmosphereComposition.toxicity > 0) && (atmosphereComposition.suffocating == 1))											{habitabilityModifier = -1;}
