+++ habitabilityModifiersTable.cpp
@@ -1,5 +1,6 @@
 //	This function corresponds to "Determining Habitability and Affinity" in "Step 7: Resources and Habitability".
 
+#include <cstdint>
 //	Constant declarations
 #include "declarations/constants/atmosphericPressureCategoryConstants.h"
 #include "declarations/constants/worldTypeConstants.h"
@@ -17,18 +18,18 @@
 #include "declarations/functions/breathableChecker.h"
 #include "declarations/functions/atmosphericPressureCategoriesTable.h"
 
-int habitabilityModifiersTable(char worldType, float atmosphereMass, float hydrographicCoverage, float atmosphericPressure, int surfaceTemperature, atmosphericComposition_t atmosphereCompositionArray)
+int8_t habitabilityModifiersTable(char worldType, float atmosphereMass, float hydrographicCoverage, float atmosphericPressure, int surfaceTemperature, atmosphericComposition_t atmosphereCompositionArray)
 {
 	//bool liquidWater = waterOceanChecker(worldType, hydrographicCoverage);
 	char climateType = climateTypeLookup(surfaceTemperature);
 	bool breathable = breathableChecker(atmosphereCompositionArray);
 	char pressureCategory = atmosphericPressureCategoriesTable(atmosphericPressure);
 
-	int habitabilityModifier = 0;
-	int breathableClimate = habitabilityModifiersBreathableClimate(climateType, breathable);
-	int breathableAtmosphere = habitabilityModifiersBreathable(pressureCategory, breathable);
-	int nonBreathableAtmosphere = habitabilityModifiersNonBreathable(pressureCategory, atmosphereCompositionArray);
-	int liquidWaterOceans = habitabilityModfiersLiquidWaterOceans(hydrographicCoverage);
+	int8_t habitabilityModifier = 0;
+	int8_t breathableClimate = habitabilityModifiersBreathableClimate(climateType, breathable);
+	int8_t breathableAtmosphere = habitabilityModifiersBreathable(pressureCategory, breathable);
+	int8_t nonBreathableAtmosphere = habitabilityModifiersNonBreathable(pressureCategory, atmosphereCompositionArray);
+	int8_t liquidWaterOceans = habitabilityModfiersLiquidWaterOceans(hydrographicCoverage);
 
 	habitabilityModifier = breathableClimate + breathableAtmosphere + nonBreathableAtmosphere + liquidWaterOceans;
